GB3-1682: PR fixes - move link list into lists directory
