+++ src/app/shared/components/lists/link-list/link-list-item/link-list-item.component.scss
@@ -1,5 +1,5 @@
-@use 'functions/helper.function' as functions;
 @use 'variables/ktzh-design-variables' as ktzh-variables;
+@use 'functions/helper.function' as functions;
 
 .link-list-item {
   margin: 0;
