+++ src/app/shared/components/lists/link-list/link-list.component.scss
@@ -1,5 +1,3 @@
-@use 'variables/ktzh-design-variables' as ktzh-variables;
-
 .link-list {
   &:not(:first-of-type) {
     margin-top: 30px;
