+++ src/app/shared/shared.module.ts
@@ -11,8 +11,8 @@ import {ExpandableListItemComponent} from './components/expandable-list-item/exp
 import {MainFooterComponent} from './components/footer/main-footer.component';
 import {ExternalLinkButtonComponent} from './components/external-link-button/external-link-button.component';
 import {HeroHeaderComponent} from './components/hero-header/hero-header.component';
-import {LinkListItemComponent} from './components/link-list/link-list-item/link-list-item.component';
-import {LinkListComponent} from './components/link-list/link-list.component';
+import {LinkListItemComponent} from './components/lists/link-list/link-list-item/link-list-item.component';
+import {LinkListComponent} from './components/lists/link-list/link-list.component';
 import {LoadingAndProcessBarComponent} from './components/loading-and-process-bar/loading-and-process-bar.component';
 import {NavbarMobileDialogComponent} from './components/navbar-mobile/navbar-mobile-dialog/navbar-mobile-dialog.component';
 import {NavbarMobileComponent} from './components/navbar-mobile/navbar-mobile.component';
